feat: add timeframe option to skill rolls (#1007)
Optional timeframes have been add to skill rolls (enabled in advanced ruleset settings).  Now have an option to specify a roll formula and units for time it takes to do a task.  There is no effect on the skill roll itself other than to add some extra information with the roll result.  If no time units are selected, a timeframe is not rolled and reported.