feat: Add encumbrance icons and modifiers (#1135)
This release adds optional encumbrance icons and modifiers.  To enable, check the display options "Display encumbrance status indicators and apply roll modifiers".  Then set the "Fraction of maximum encumbrance when actor becomes encumbered" value in advanced ruleset settings.  Please note that this value multiplies the prior setting for the maximum encumbrance formula (max carry weight) to trigger the encumbered status.  For example, most Cepheus derived rules trigger at one third the maximum weight.  For such cases, the fraction of maximum encumbrance should be 0.33.  The other relevant encumbered setting is the DM to apply to physical checks while encumbered.  Typically -1 or -2.  These DM's are applied to ANY roll that uses a physical characteristic (STR, DEX, END).  The encumbered condition adjusts the base characteristic modifier using an active effect.  So, the roll information will display the net modifier in the details.  For example, STR(-1) will show for a character with a STR of 7(0) that is encumbered where the DM is -1.  This is still a beta feature.