feat: Improvements to roll dialog and chat message (#1172)
Are you a GM who wants more control over rolls?  Are your players complaining that their roll is missing a bonus? Don't trust all the behind-the-scenes magic in Twodsix?  Well this update is for you!

This update greatly expands the roll dialog so that all relevant settings can now be modified / overridden (with the exception of characteristic modifier values).  Also, the resulting chat message now includes more information and the description order should match the roll formula.