feat: add armament ammo counter (#1261)
This update adds fields to the armament component sheet so that ammunition useage may be tracked.  Two new fields under the "Ammunition" header appear on the armament sheet as "# / #" where the left number is the available ammo and the right is the maximum that can be held by the component.  Also, the component lists for the ship sheet will show available ammo count and +/- toggle buttons.  These features will only appear if the max ammo is greater than zero.  Otherwise available and total quantity are shown.